+++ lce/src/main/java/com/laimiux/lce/CT.kt
@@ -8,6 +8,20 @@ interface CT<out C> {
     companion object {
         fun <T> content(content: T): CT<T> = Type.Content(content)
         fun error(error: Throwable): CT<Nothing> = Type.Error(error)
+
+        /**
+         * Returns [onNull] if [content] is null, otherwise returns [CT.content].
+         */
+        inline fun <L, C : Any, E> fromNullable(
+            content: C?,
+            crossinline onNull: () -> CT<C>
+        ): CT<C> {
+            return if (content == null) {
+                onNull()
+            } else {
+                content(content)
+            }
+        }
     }
 
     fun isContent(): Boolean